requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.58.2
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-etc==1.4.0
django-google-translate==1.1
django-heroku==0.3.1
django-hitcount==1.3.5
django-js-asset==2.1.0
django-social-share==2.3.0
django-taggit==5.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==4.9.3
msgpack==1.0.7
oauthlib==3.2.2
openai==1.3.8
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
wikipedia==1.4.0
